refactor: move rpc's commons to common-meta
