+++ src/common/meta/src/lib.rs
@@ -15,9 +15,8 @@
 pub mod error;
 pub mod instruction;
 pub mod peer;
-pub mod router;
+pub mod rpc;
 pub mod table_name;
-pub mod util;
 
 pub type ClusterId = u64;
 pub type DatanodeId = u64;
