+++ src/common/meta/src/rpc/router.rs
@@ -26,8 +26,8 @@ use table::metadata::RawTableInfo;
 
 use crate::error::{self, Result};
 use crate::peer::Peer;
+use crate::rpc::util;
 use crate::table_name::TableName;
-use crate::util;
 
 #[derive(Debug, Clone)]
 pub struct CreateRequest<'a> {
