+++ src/frontend/src/instance/distributed.rs
@@ -30,10 +30,11 @@ use client::Database;
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
 use common_catalog::format_full_table_name;
 use common_error::prelude::BoxedError;
-use common_meta::router::{
+use common_meta::rpc::router::{
     CreateRequest as MetaCreateRequest, DeleteRequest as MetaDeleteRequest,
     Partition as MetaPartition, RouteRequest, RouteResponse,
 };
+use common_meta::rpc::store::CompareAndPutRequest;
 use common_meta::table_name::TableName;
 use common_query::Output;
 use common_telemetry::debug;
@@ -42,7 +43,6 @@ use datanode::sql::SqlHandler;
 use datatypes::prelude::ConcreteDataType;
 use datatypes::schema::RawSchema;
 use meta_client::client::MetaClient;
-use meta_client::rpc::CompareAndPutRequest;
 use partition::manager::PartitionInfo;
 use partition::partition::{PartitionBound, PartitionDef};
 use query::error::QueryExecutionSnafu;
