+++ src/meta-client/src/lib.rs
@@ -18,7 +18,6 @@ pub mod client;
 pub mod error;
 #[cfg(test)]
 mod mocks;
-pub mod rpc;
 
 // Options for meta client in datanode instance.
 #[derive(Clone, Debug, Serialize, Deserialize)]
